requirements.txt
fastapi[standard]
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
google-auth 
google-auth-oauthlib 
google-auth-httplib2